chore: revise template scripts
save time to create new components
