+++ .gitignore
@@ -19,3 +19,7 @@ coverage/
 
 # files related to the project (Mood board etc.)
 resources/
+
+# Uncomment once cloned
+## template scripts
+#template*.js
