+++ src/blocks/TemplateBlock.js
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+
+const TemplateBlock = styled.div``;
+
+export default TemplateBlock;
