--- src/components/Template.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-
-const Template = () => {
-  return <div>Welcome to Next.js!</div>;
-};
-
-Template.propTypes = {};
-
-export default Template;
