+++ src/components/TemplateComponent.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+import TemplateBlock from 'src/blocks/TemplateBlock';
+
+const TemplateComponent = () => {
+  return <TemplateBlock>Welcome to Next.js!</TemplateBlock>;
+};
+
+TemplateComponent.propTypes = {};
+
+export default TemplateComponent;
