+++ src/elements/TemplateElement.js
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+
+const TemplateElement = styled.div``;
+
+export default TemplateElement;
