--- src/styledComponents/StyledComponent.js
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-import PropTypes from 'prop-types';
-
-const StyledComponent = styled.div`
-  width: 100%;
-`;
-
-StyledComponent.propTypes = {};
-
-export default StyledComponent;
